fix(#501): CalculateIoU の整数乗算オーバーフローを修正
CodeQL警告 "Possible overflow: result of integer multiplication cast to float" を修正。
int * int の結果を float にキャストする前にオーバーフローする可能性があったため、
乗算前に float にキャストするように変更。

+++ Baketa.Infrastructure/Processing/Strategies/TextChangeDetectionStageStrategy.cs
@@ -365,8 +365,8 @@ public partial class TextChangeDetectionStageStrategy : IProcessingStageStrategy
         if (intersection.IsEmpty)
             return 0f;
 
-        float intersectionArea = intersection.Width * intersection.Height;
-        float unionArea = (a.Width * a.Height) + (b.Width * b.Height) - intersectionArea;
+        float intersectionArea = (float)intersection.Width * intersection.Height;
+        float unionArea = ((float)a.Width * a.Height) + ((float)b.Width * b.Height) - intersectionArea;
 
         return unionArea > 0 ? intersectionArea / unionArea : 0f;
     }
